requirements.txt
google-api-python-client
google-cloud-BigQuery
google-cloud-logging